+++ src/app/Fake.Core.Trace/TraceListener.fs
@@ -7,6 +7,8 @@ open System
 type KnownTags =
     | Task of name:string
     | Target of name:string
+    | FinalTarget of name:string
+    | FailureTarget of name:string
     | Compilation of compiler:string
     | TestSuite of suiteName:string
     | Test of testName:string
@@ -15,6 +17,8 @@ type KnownTags =
         match x with
         | Task n
         | Target n
+        | FinalTarget n
+        | FailureTarget n
         | Compilation n
         | TestSuite n
         | Test n
@@ -23,6 +27,8 @@ type KnownTags =
         match x with
         | Task _ -> "task"
         | Target _ -> "target"
+        | FinalTarget _ -> "final target"
+        | FailureTarget _ -> "failure target"
         | Compilation _ -> "compilation"
         | TestSuite _ -> "testsuite"
         | Test _ -> "test"
@@ -100,7 +106,7 @@ module TestStatus =
             TestStatus.Failed (f message, f details, Some (f expected, f actual))
         | TestStatus.Failed (message, details, None) ->
             TestStatus.Failed (f message, f details, None)
-        | _ -> t        
+        | _ -> t
 
 [<RequireQualifiedAccess>]
 type TagStatus =
@@ -110,14 +116,15 @@ type TagStatus =
 
 /// Defines Tracing information for TraceListeners
 [<RequireQualifiedAccess>]
-type TraceData = 
+type TraceData =
     | ImportData of typ:ImportData * path:string
     | BuildNumber of text:string
     | ImportantMessage of text:string
     | ErrorMessage of text:string
     | LogMessage of text:string * newLine:bool
     | TraceMessage of text:string * newLine:bool
-    | OpenTag of KnownTags * description:string
+    /// Happens when a tag (Task, Target, Test, ...) has started.
+    | OpenTag of KnownTags * description:string option
     | TestStatus of testName:string * status:TestStatus
     | TestOutput of testName:string * out:string * err:string
     | CloseTag of KnownTags * time:TimeSpan * TagStatus
@@ -152,7 +159,7 @@ type TraceData =
 module TraceData =
     let inline mapMessage f (t:TraceData) =
         match t with
-        | TraceData.ImportantMessage text -> TraceData.ImportantMessage (f text) 
+        | TraceData.ImportantMessage text -> TraceData.ImportantMessage (f text)
         | TraceData.ErrorMessage text -> TraceData.ErrorMessage (f text)
         | TraceData.LogMessage (text, d) -> TraceData.LogMessage (f text, d)
         | TraceData.TraceMessage (text, d) -> TraceData.TraceMessage (f text, d)
@@ -166,12 +173,12 @@ module TraceData =
         mapMessage (repl oldString replacement) t
 
 /// Defines a TraceListener interface
-type ITraceListener = 
+type ITraceListener =
     abstract Write : TraceData -> unit
 
 module ConsoleWriter =
-    
-    let write toStdErr color newLine text = 
+
+    let write toStdErr color newLine text =
         let curColor = Console.ForegroundColor
         try
           if curColor <> color then Console.ForegroundColor <- color
@@ -219,7 +226,7 @@ module ConsoleWriter =
         printer "\x1b[%sm%s\x1b[0m" codeStr text
 
     /// A default color map which maps TracePriorities to ConsoleColors
-    let colorMap traceData = 
+    let colorMap traceData =
         match traceData with
         | TraceData.ImportantMessage _ -> ConsoleColor.Yellow
         | TraceData.ErrorMessage _ -> ConsoleColor.Red
@@ -234,7 +241,7 @@ module ConsoleWriter =
 type ConsoleTraceListener(importantMessagesToStdErr, colorMap, ansiColor) =
     interface ITraceListener with
         /// Writes the given message to the Console.
-        member __.Write msg = 
+        member __.Write msg =
             let color = colorMap msg
             let write = if ansiColor then ConsoleWriter.writeAnsiColor else ConsoleWriter.write
             match msg with
@@ -242,14 +249,17 @@ type ConsoleTraceListener(importantMessagesToStdErr, colorMap, ansiColor) =
                 write importantMessagesToStdErr color true text
             | TraceData.LogMessage(text, newLine) | TraceData.TraceMessage(text, newLine) ->
                 write false color newLine text
-            | TraceData.OpenTag(KnownTags.Target _ as tag, description) ->
-                write false color true (sprintf "Starting %s '%s'" tag.Type tag.Name)
-                if not (isNull description) then
-                    let msg = TraceData.TraceMessage("", true)
-                    let color2 = colorMap msg
-                    write false color2 true description
-            | TraceData.OpenTag (tag, descr) ->
-                write false color true (sprintf "Starting %s '%s': %s" tag.Type tag.Name descr)
+            | TraceData.OpenTag(KnownTags.Target _ as tag, description)
+            | TraceData.OpenTag(KnownTags.FailureTarget _ as tag, description)
+            | TraceData.OpenTag(KnownTags.FinalTarget _ as tag, description) ->
+                let color2 = colorMap (TraceData.TraceMessage("", true))
+                match description with
+                | Some d -> write false color2 true (sprintf "Starting %s '%s': %s" tag.Type tag.Name d)
+                | _ -> write false color2 true (sprintf "Starting %s '%s'" tag.Type tag.Name)                
+            | TraceData.OpenTag (tag, description) ->
+                match description with
+                | Some d -> write false color true (sprintf "Starting %s '%s': %s" tag.Type tag.Name d)
+                | _ -> write false color true (sprintf "Starting %s '%s'" tag.Type tag.Name)                
             | TraceData.CloseTag (tag, time, status) ->
                 write false color true (sprintf "Finished (%A) '%s' in %O" status tag.Name time)
             | TraceData.ImportData (typ, path) ->
@@ -269,7 +279,7 @@ type TraceSecret =
 
 module TraceSecrets =
     let private traceSecretsVar = "Fake.Core.Trace.TraceSecrets"
-    let private getTraceSecrets, _, (setTraceSecrets:TraceSecret list -> unit) = 
+    let private getTraceSecrets, _, (setTraceSecrets:TraceSecret list -> unit) =
         Fake.Core.FakeVar.define traceSecretsVar
 
     let getAll () =
@@ -289,8 +299,8 @@ module TraceSecrets =
 
 module CoreTracing =
     // If we write the stderr on those build servers the build will fail.
-    let importantMessagesToStdErr = 
-        let buildServer = BuildServer.buildServer 
+    let importantMessagesToStdErr =
+        let buildServer = BuildServer.buildServer
         buildServer <> CCNet && buildServer <> AppVeyor && buildServer <> TeamCity && buildServer <> TeamFoundation
 
     /// The default TraceListener for Console.
@@ -301,7 +311,7 @@ module CoreTracing =
     /// A List with all registered listeners
 
     let private traceListenersVar = "Fake.Core.Trace.TraceListeners"
-    let private getTraceListeners, _, (setTraceListenersPrivate:ITraceListener list -> unit) = 
+    let private getTraceListeners, _, (setTraceListenersPrivate:ITraceListener list -> unit) =
         Fake.Core.FakeVar.define traceListenersVar
 
     let areListenersSet () =