+++ src/template/fake-template/Content/fake.tool.sh
@@ -3,10 +3,25 @@
 set -eu
 set -o pipefail
 
-TOOL_PATH=(ToolPath)
+# liberated from https://stackoverflow.com/a/18443300/433393
+realpath() {
+  OURPWD=$PWD
+  cd "$(dirname "$1")"
+  LINK=$(readlink "$(basename "$1")")
+  while [ "$LINK" ]; do
+    cd "$(dirname "$LINK")"
+    LINK=$(readlink "$(basename "$1")")
+  done
+  REALPATH="$PWD/$(basename "$1")"
+  cd "$OURPWD"
+  echo "$REALPATH"
+}
 
-if ! [ -e $TOOL_PATH/fake ] 
+TOOL_PATH=$(realpath (ToolPath))
+FAKE="$TOOL_PATH"/fake
+
+if ! [ -e "$FAKE" ]
 then
   dotnet tool install fake-cli --tool-path $TOOL_PATH --version (version)
 fi
-$TOOL_PATH/fake "$@"
+"$FAKE" "$@"